refactor: change logic
