+++ package.json
@@ -35,20 +35,37 @@
   },
   "dependencies": {
     "@babel/core": "^7.9.0",
+    "@babel/traverse": "^7.10.1",
     "awesome-typescript-loader": "^5.2.1",
     "babel-loader": "^8.1.0",
-    "babel-preset-alexios": "^0.2.1",
+    "babel-preset-alexios": "^0.2.3",
+    "body-parser": "^1.18.3",
     "chalk": "^4.0.0",
-    "commands": "^0.0.7",
+    "chokidar": "^3.0.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^6.0.2",
+    "css-loader": "^3.5.3",
+    "debug": "^3.1.0",
     "file-loader": "^6.0.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "fs-extra": "^9.0.0",
+    "glob": "^7.1.2",
+    "happypack": "^5.0.1",
     "html-webpack-plugin": "^4.2.0",
+    "less-loader": "^6.1.0",
+    "mini-css-extract-plugin": "^0.9.0",
+    "nodemon": "^2.0.4",
+    "path-to-regexp": "^2.2.1",
     "react": "^16.13.1",
+    "react-app-polyfill": "^1.0.6",
     "react-dom": "^16.13.1",
+    "sass-loader": "^8.0.2",
+    "style-loader": "^1.2.1",
     "typescript": "^3.8.3",
+    "url-loader": "^4.1.0",
     "watch": "^1.0.2",
     "webpack": "^4.42.1",
+    "webpack-bundle-analyzer": "^3.8.0",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
     "webpack-progress-bar": "^1.2.1",
