--- src/babel.js
@@ -1,3 +1,0 @@
-require("@babel/register")({
-  presets: ["@babel/preset-env"],
-});
