--- src/webpack/confg/babel-loader.js
@@ -1,8 +1,0 @@
-const babelLoader = {
-  loader: "babel-loader",
-  options: {
-    presets: ["alexios"],
-  },
-};
-
-export default babelLoader;
