--- src/webpack/main/build-generator.js
@@ -1,13 +1,0 @@
-/**
- * 打包配置
- *
- * @param {object} props
- *
- * @returns {object} webpackConfig
- */
-
-const buildGenerator = () => {
-  // TODO: 项目打包配置!!
-};
-
-export default buildGenerator;
