--- src/webpack/main/index.js
@@ -1,1 +1,0 @@
-export { default as webpackDev } from "./dev-generator.js";
