--- templates/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "esnext",
-    "target": "esnext",
-    "lib": ["esnext", "dom"],
-    "sourceMap": true,
-    "jsx": "react",
-    "allowSyntheticDefaultImports": true,
-    "moduleResolution": "node",
-    "allowJs": true,
-    "skipLibCheck": true
-  },
-  "exclude": ["node_modules", "dist"]
-}
